content/base/test/test_bug598877.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=598877
-->
<head>
  <title>Test for Bug 598877</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=598877">Mozilla Bug 598877</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 598877 **/

var attr = document.createAttribute("foo");
attr.value = "bar";
ok(attr.firstChild, "attr.firstChild should be a text node");
is(attr.firstChild.nodeValue, "bar", "the text node value should be 'bar'");

</script>
</pre>
</body>
</html>
